backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.0
werkzeug==2.3.7